requirements.txt
flask
pandas
pillow
gspread
oauth2client
fastapi
uvicorn
python-dotenv
cloudinary
slowapi
locust
